+++ vantage6-ui/src/app/pages/analyze/task/create/task-create.component.ts
@@ -345,7 +345,7 @@ export class TaskCreateComponent implements OnInit, OnDestroy, AfterViewInit {
       Object.keys(this.parameterForm.controls).forEach((control) => {
         if (control === arg.name) {
           const value = this.parameterForm.get(control)?.value;
-          if (arg.has_default_value && value === null) {
+          if (arg.is_frontend_only || (arg.has_default_value && value === null)) {
             return; // note that within .forEach, return is like continue
           } else if (arg.type === ArgumentType.Json) {
             kwargs[arg.name] = JSON.parse(value);